requirements.txt
azure-common==1.1.26
azure-core==1.11.0
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
idna==2.10
isodate==0.6.0
msal==1.9.0
msal-extensions==0.3.0
msrest==0.6.21
mysql-connector-python==8.0.23
oauthlib==3.1.0
pandas==1.1.5
portalocker==1.7.1
protobuf==3.15.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
tqdm==4.57.0
urllib3==1.26.3
